| Refactor TransmissionRouter location
