+++ sample/src/main/java/com/trendyol/transmission/SampleViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.trendyol.transmission.effect.RouterEffect
 import com.trendyol.transmission.features.colorpicker.ColorPickerTransformer
 import com.trendyol.transmission.features.input.InputTransformer
+import com.trendyol.transmission.router.TransmissionRouter
 import com.trendyol.transmission.router.onEach
 import com.trendyol.transmission.router.toState
 import com.trendyol.transmission.ui.ColorPickerUiState
