+++ sample/src/main/java/com/trendyol/transmission/features/FeaturesModule.kt
@@ -1,6 +1,6 @@
 package com.trendyol.transmission.features
 
-import com.trendyol.transmission.TransmissionRouter
+import com.trendyol.transmission.router.TransmissionRouter
 import com.trendyol.transmission.features.colorpicker.ColorPickerTransformer
 import com.trendyol.transmission.features.input.InputTransformer
 import com.trendyol.transmission.features.multioutput.MultiOutputTransformer
