+++ transmission/src/main/java/com/trendyol/transmission/effect/RouterEffect.kt
@@ -1,7 +1,7 @@
 package com.trendyol.transmission.effect
 
 import com.trendyol.transmission.Transmission
-import com.trendyol.transmission.TransmissionRouter
+import com.trendyol.transmission.router.TransmissionRouter
 import com.trendyol.transmission.transformer.Transformer
 
 /**
