+++ transmission/src/main/java/com/trendyol/transmission/transformer/dataholder/DataHolderExt.kt
@@ -1,7 +1,7 @@
 package com.trendyol.transmission.transformer.dataholder
 
 import com.trendyol.transmission.Transmission
-import com.trendyol.transmission.TransmissionRouter
+import com.trendyol.transmission.router.TransmissionRouter
 import com.trendyol.transmission.transformer.Transformer
 import com.trendyol.transmission.transformer.request.Contract
 import com.trendyol.transmission.transformer.request.RequestHandler
